\name{soiltexturetransformation-package}
\alias{soiltexturetransformation-package}
\alias{soiltexturetransformation}
\docType{package}
\title{Functions for soil texture transformation}
\description{A set of R functions designed to transform soil textures data.}
\details{
\tabular{ll}{Package: \tab soiltexturetransformation\cr
Version: \tab 1.0.0\cr
Date: \tab 2014-01-09\cr
Title: \tab Functions for soil texture transformation\cr
Author: \tab Wei Shangguan\cr
Maintainer: \tab Julien MOEYS <jules_m78-soiltexture@yahoo.fr>\cr
Depends: \tab R (>= 3.0.2), soiltexture, drc\cr
Suggests: \tab \cr
License: \tab AGPL (>=3)\cr
URL: \tab http://soiltexture.r-forge.r-project.org/\cr}
}
\author{Wei Shangguan}

\keyword{ package }


